After discussing with your tutor Jack, there seems to be some students who have increased speed of convergence significantly by effectively multiplying the loss by 2048. This is somewhat equivalent to increasing the learning rate (although it also helps more than this because it avoids float32 underflow and vanishing gradient issues). It is okay if you want to make this adjustment to the network as convergence time impacts how quickly you can test changes in your code -- but do it correctly!
Also, as this is not an architectural change it may help you get better PSNR numbers early in training but does not count for the modification you need to make.

The root cause of finding this is that some students have modified their ray_attenuation function like this:

ray_attenuations = ... calculations that compute the average attenuation over the distance in the ray ... # shape = (2048,)
return tf.expand_dims(ray_attenuations, axis=-1) # shape = (2048,1)
The above code produces incorrect values for the loss because the gradient is now computed 2048 times. Tensorflow does not report an error but this is extremely inefficient (and violates the contract of the gradient API). To get the same result you can instead have your ray_attenuation function return normally:

ray_attenuations = ... calculations that compute the average attenuation over the distance in the ray ... # shape = (2048,)
return ray_attenuations # shape = (2048,)

and then modify the loss computation:
loss = tf.keras.losses.MSE(projection, predicted_attenuation) * 2048

which will give you the same result and is valid.